+++ src/webviews/lwc-ui-panel.ts
@@ -120,6 +120,8 @@ export class LwcUiPanel {
       "s-prompt-input": "Prompt Input",
       "s-command-execution": "Command Execution",
       "s-pipeline": "DevOps Pipeline",
+      "s-pipeline-config": "Pipeline Settings",
+      "s-extension-config": "Extension Settings",
     };
     const panelTitle = lwcDefinitions[this.lwcId] || "SFDX Hardis";
     this.panel.title = panelTitle;
@@ -439,6 +441,10 @@ export class LwcUiPanel {
   private async handleUpdateVsCodeSfdxHardisConfiguration(data: { configKey: string; value: any }): Promise<void> {
     try {
       const config = vscode.workspace.getConfiguration("vsCodeSfdxHardis");
+      if (data.configKey.startsWith("vsCodeSfdxHardis.")) {
+        data.configKey = data.configKey.replace("vsCodeSfdxHardis.", "");
+      }
+      // Update the configuration value
       await config.update(data.configKey, data.value, vscode.ConfigurationTarget.Global);
       vscode.window.showInformationMessage(`VsCode configuration '${data.configKey}' updated with value: ${data.value}`);
     } catch (error) {
@@ -515,7 +521,7 @@ export class LwcUiPanel {
           #app { width: 100%; height: 100vh; }
         </style>
       </head>
-      <body class="slds-scope">
+      <body class="slds-scope slds-theme_default">
         <div id="app" data-lwc-id="${this.lwcId}" data-init-data="${initDataJson}"></div>
         
         <script>